feat: use mdx as part of pages
