--- content/avdelinger/index.md
@@ -1,4 +1,0 @@
----
-title: Lokasjoner
-entry: trondheim.md
----
