--- content/avdelinger/oslo.md
@@ -1,80 +1,0 @@
----
-title: Oslo
-order: 1
----
-
-
-## Spaces Kvadraturen
-Vi holder til, så lenge vi ikke er ute hos våre kunder, hos
-[Spaces](https://g.page/Tollbugata-8-4308?share) i Kvadraturen.
-
-### WiFi
-Når du jobber på kontoret benytter du det trådløse nettet med SSID
-**Spaces**. Du logger inn med din Variant-epost.
-
-### Sykkelparkering
-Spaces har både sykkelparkering og garderobe. Det er litt knotete å få sykkelen ned (må bruke heisen). 
-Be resepsjonen på Spaces om å vise deg rutinen og hvor garderoben ligger. 
-
-## Faktura
-Hvis du får en faktura som skal betales av Variant Oslo sendes den til
-faktura.oslo@variant.no. 
-
-## Timeføring
-
-Timene skal leveres ukentlig. Fristen er mandag morgen kl 9.00 den påfølgende
-uken. Timene føres i [Harvest](https://variantosloas.harvestapp.com/). Herfra kan
-man også laste ned Apps for både Mac og telefoner.
-
-## Helse, miljø og sikkerhet
-
-### Verneombud
-
-Selskapet skal legge til rette for at det velges et verneombud. Verneombudet velges for to år om gangen og velges på variantdagen i mars. Ledelsen kan ikke delta i valget eller velges som verneombud. Se lokale sider for hvem som verneombud i de ulike byene. Verneombud fra Mars 2022 i Oslo er Jonas Lillevold og kontaktinformasjon finnes på hjemmesiden vår og i internsystemer.
-
-### Lokalt HMS-arbeid
-
-Variant er et konsulentselskap, der hvor mange ansatte tilbringer store deler av
-arbeidstiden ute i kundenes lokaler. HMS-håndboka dekker proaktivt arbeid med
-lokaler, men vi avgrenser dette til vårt kontor på Spaces.
-
-- Renhold i våre lokaler gjennomføres i regi av Spaces
-- Førstehjelpsutstyr ligger under vasken i fellesområdet og er til enhver tid komplett i henhold til [retningslinjer gitt av Arbeidstilsynet](https://www.arbeidstilsynet.no/tema/personlig-verneutstyr/forstehjelpsutstyr/)
-- Ulykker som medfører alvorlig personskade skal straks meldes til lege/ambulanse, politi og Arbeidstilsynets lokalkontor. Innmelding skjer av daglig leder
-- Gårdeier sørger for avfallssytemene i bygget vi leier kontorplass i. Det er kildesorteringssystem med instruks, godt synlig i fellesområdet. Gode rutiner skal bidra til å redusere svinn og avfall
-
-Hvis ansatte opplever feil, mangler eller farlige situasjoner ute hos kunde
-ønsker vi at dette forsøkes varslet i henhold til kundens HMS-rutiner. Skulle
-dette ikke føre fram vil vi ta imot og håndtere saken i Variant-kontekst.
-
-### Branninstruks
-
-Gjør deg kjent med følgende hos Spaces Kvadraturen:
-
-- Rømningsveiene
-- Nærmeste brannmelder
-- Nærmeste slokkemiddel
-
-Ved brannalarm, forlat bygningen gjennom rømningsveiene og gå deretter til [samlingsplassen over gaten på hjørnet UTENFOR ved Pascal.](https://g.page/pascal-avd-tollbugata) Hvis brann oppstår uten at automatisk alarm er utløst, trykk på brannmelder hvis mulig og gå deretter ut av bygningen og til samlingsplassen.
-
-Vinduer og døren til kontoret må lukkes. Om mulig, forsøk å slokke brann med enten pulverapparat eller slange, men tenk på egen sikkerhet. Heis skal IKKE benyttes som rømningsvei.
-
-Nødnummer til brannvesen er 110.
-
-Rømningsveier og brannslykkingsutstyr er godt merket. Branninnstruks og redningsplan finnes som oppslag i kontorlokalet og gjennomgås med alle nyansatte. Brannøvelse gjennomføres i regi av Spaces. Vedlikehold av brannvarslere og slukkeustyr utføres av Spaces.
-
-### Kildesortering
-
-Det er kildesortering hos Spaces, både i fellesarealet og i søppelrommet i kjelleren. Gjør deg kjent med hvordan kildesortering fungerer i Oslo [på hjemmesiden til Oslo Kommune](https://www.oslo.kommune.no/avfall-og-gjenvinning/hvordan-kildesortere-i-oslo/).
-
-## Personalledere
-
-For tiden har følgende varianter personalansvar:
-
-- Linn
-- Lars
-- Nikolai
-
-Relasjonen mellom deg og personallederen er basert på coaching. Se også
-[listen over hvem som har lederansvar for hvem](https://varianttrh.sharepoint.com/:x:/g/EcciKCrwQFVKi9arVFV9WLEBxh0lqSzlfM7n_nrpkIuDtw?e=gR7JSm).
-
