+++ package.json
@@ -11,20 +11,20 @@
   "description": "Variant's handbook-v2",
   "devDependencies": {
     "@types/mdx": "^2.0.1",
-    "@types/node": "^17.0.7",
+    "@types/node": "^17.0.8",
     "@types/react": "^17.0.38",
-    "@typescript-eslint/parser": "^5.8.1",
+    "@typescript-eslint/parser": "^5.9.0",
     "eslint": "^8.6.0",
     "eslint-plugin-react-hooks": "^4.3.0",
     "husky": "^7.0.4",
-    "lint-staged": "^12.1.5",
+    "lint-staged": "^12.1.7",
     "next-compose-plugins": "^2.2.1",
     "prettier": "^2.5.1",
     "typescript": "^4.5.4"
   },
   "dependencies": {
     "@mdx-js/loader": "^2.0.0-rc.2",
-    "@mdx-js/mdx": "^2.0.0-rc.2",
+    "@mdx-js/mdx": "2.0.0-next.9",
     "@mdx-js/react": "^2.0.0-rc.2",
     "@next/mdx": "^12.0.7",
     "@variant/md-section": "^1.0.0",
@@ -37,7 +37,11 @@
     "react-dom": "17.0.2",
     "react-hotkeys-hook": "^3.4.4",
     "react-select": "^5.2.1",
+    "remark-frontmatter": "^4.0.1",
+    "remark-gfm": "^3.0.1",
+    "remark-mdx-frontmatter": "^1.1.1",
     "remark-mdx-toc": "^0.1.0",
+    "slugify": "^1.6.5",
     "swr": "^1.1.2"
   },
   "scripts": {
