+++ pages/avdelinger/index.mdx
@@ -1,0 +1,7 @@
+---
+title: Lokasjoner
+---
+
+import Trondheim from "./trondheim.mdx";
+
+<Trondheim />
