+++ pages/information.mdx
@@ -412,3 +412,7 @@ spør du i #helpdesk 🤓
   bruk av bonuspoeng eller andre fordelsprogam.
 - [Fordelsdokumentasjonen](https://fordelsdokumentasjon.variant.no) Hit laster
   du opp dokumentasjon på ovennevnt bruk.
+
+import { HandbookLayout } from "src/handbook-layout";
+
+export default (props) => <HandbookLayout meta={meta} toc={toc} {...props} />;
