+++ pages/quality_manual.mdx
@@ -52,9 +52,9 @@ bemanningsprosessen sammen med salgsprosessen.
 Lønnsomhet sikres ved å ha høy kvalitet i alt salgs- og tilbudsarbeid slik at vi
 vinner nok salg. Variants tilbud skal alltid levere over kundens forventninger.
 
-Salgsprosessen i Variant samkjøres på tvers av alle selskap i konsernet og eies av 
-konsernets salgsdirektør Anders Hammervold. De ulike selskapene har likefullt et ansvar for 
-salg av sine konsulenter. 
+Salgsprosessen i Variant samkjøres på tvers av alle selskap i konsernet og eies
+av konsernets salgsdirektør Anders Hammervold. De ulike selskapene har likefullt
+et ansvar for salg av sine konsulenter.
 
 ### Salgsprosess
 
@@ -185,9 +185,9 @@ helse, miljø og sikkerhet som ikke løses direkte.
 ### Verneombud
 
 Selskapet skal legge til rette for at det velges et verneombud. Verneombudet
-velges for to år om gangen og velges på variantdagen i mars.  Ledelsen kan ikke
-delta i valget eller velges som verneombud. På lokasjonssidene 
-kan du se hvem som er valgt til for hver lokasjon.
+velges for to år om gangen og velges på variantdagen i mars. Ledelsen kan ikke
+delta i valget eller velges som verneombud. På lokasjonssidene kan du se hvem
+som er valgt til for hver lokasjon.
 
 Verneombud har ansvar for opplæring av ansatte og oppdatering av aktuelle
 rutiner. Ved nyansettelser skal verneombud ha opplæring og vernerunde. Skjer
@@ -502,16 +502,16 @@ Hver ansatt knyttes til en fadder og en personalleder.
 
 #### Fadder
 
-Den som har ansvaret for at oppstarten går så bra som mulig kaller vi fadder.  Fadderens overordnede ansvar er å gi deg
-en så god oppstart som mulig. Arbeidet starter ganske raskt etter signering og
-lenge før du starter hos oss. Variant.
+Den som har ansvaret for at oppstarten går så bra som mulig kaller vi fadder.
+Fadderens overordnede ansvar er å gi deg en så god oppstart som mulig. Arbeidet
+starter ganske raskt etter signering og lenge før du starter hos oss. Variant.
 
 En fadder skal være kontaktpunktet for alle praktiske spørsmål om hvordan ting
 gjøres og løses i Variant. I tillegg skal fadder følge opp den ansatte i starten
 med tanke på trivsel og motivasjon.
 
 Det er ingen formell slutt på fadderperioden. Så lenge den ansatte ser verdien
-av å bruke sin fadder er det åpent for det. 
+av å bruke sin fadder er det åpent for det.
 
 Se lokasjonsspesifikk side for å finne hvem som er fadder for ditt sted.
 
@@ -527,16 +527,17 @@ ansvar for å jevnlig følge opp deg. I tillegg vil personalleder ta initiativ t
 
 I Variant praktiserer vi fritt ledervalg. Det vil si at selskapet foreslår en
 personalleder for deg ved oppstart. I denne prosessen skal Variant avklare
-valget med deg. Hver høst fasiliterer selskapet et nytt ledervalg. Her kan den ansatte hvis hen 
-ønsker velge en ny leder. Dette kan være av både karriæremessige eller personlige grunner. 
-Det stilles ingen spørsmål rundt dette valget. Det er slett ikke noen forventning om at alle skal 
-bytte leder heller. 
+valget med deg. Hver høst fasiliterer selskapet et nytt ledervalg. Her kan den
+ansatte hvis hen ønsker velge en ny leder. Dette kan være av både
+karriæremessige eller personlige grunner. Det stilles ingen spørsmål rundt dette
+valget. Det er slett ikke noen forventning om at alle skal bytte leder heller.
 
-En personalleder må være en del av selskapets ledelse, være motivert for å
-være personalleder, selv ha valgt å være det, og ha erfaring som coach og med
+En personalleder må være en del av selskapets ledelse, være motivert for å være
+personalleder, selv ha valgt å være det, og ha erfaring som coach og med
 coaching.
 
-Se lokasjonsspesifikk side for å finne de ulike personalederne som gjelder for ditt sted.
+Se lokasjonsspesifikk side for å finne de ulike personalederne som gjelder for
+ditt sted.
 
 ### Styrkesamtale
 
@@ -558,3 +559,6 @@ Basert på dine målsetninger og ønsket vekst blir dere enige om hvordan dere s
 jobbe videre. Din personalleder vil støtte deg og coache deg i tråd med det dere
 er enige om.
 
+import { HandbookLayout } from "src/handbook-layout";
+
+export default (props) => <HandbookLayout meta={meta} toc={toc} {...props} />;
