+++ src/handbook-layout/index.tsx
@@ -2,33 +2,20 @@ import React from "react";
 import Layout from "src/layout";
 import Book from "src/components/book";
 import { MDXProvider } from "@mdx-js/react";
-import { slugify } from "src/util";
+import slugify from "slugify";
 
 type HandbookProps = React.PropsWithChildren<{
-  meta: {
-    title: string;
-  };
+  meta: Metadata;
   toc: Toc;
 }>;
 
 function LinkableH2({ children, ...props }: JSX.IntrinsicElements["h2"]) {
-  console.log("inside here");
   const textContent = getNodeText(children);
   return (
-    <h2 {...props} id={slugify(textContent)}>
+    <h2 {...props} id={slugify(textContent, { lower: false })}>
       {children}
     </h2>
   );
-}
-
-function getNodeText(node: React.ReactNode): string {
-  if (typeof node == "string") return node;
-  if (typeof node == "number") return String(node);
-  if (node == null) return "";
-  if (Array.isArray(node)) return node.map(getNodeText).join("");
-  console.log(node);
-  return "";
-  // return getNodeText(node);
 }
 
 const components = {
@@ -37,7 +24,7 @@ const components = {
 
 export function HandbookLayout(props: HandbookProps) {
   const subHeadings = props.toc[0].children.map((c) => c.value);
-  console.log(props);
+
   return (
     <MDXProvider components={components}>
       <Layout subHeadings={subHeadings}>
@@ -45,4 +32,16 @@ export function HandbookLayout(props: HandbookProps) {
       </Layout>
     </MDXProvider>
   );
+}
+
+function getNodeText(node: React.ReactNode): string {
+  if (typeof node == "string") return node;
+  if (typeof node == "number") return String(node);
+  if (typeof node == "boolean") return "";
+  if (node == null) return "";
+  if (Array.isArray(node)) return node.map(getNodeText).join("");
+  if ("props" in node) {
+    return getNodeText(node.props.children);
+  }
+  return "";
 }
