+++ src/types/metadata.d.ts
@@ -1,0 +1,3 @@
+type Metadata = {
+  title: string;
+};
