+++ src/types/toc.d.ts
@@ -1,0 +1,6 @@
+type TocItem = {
+  depth: number;
+  value: string;
+  children: TocItem[];
+};
+type Toc = TocItem[];
