--- src/util.ts
@@ -1,3 +1,0 @@
-export function slugify(heading: string) {
-  return heading.replace(/[ -]+/g, "-").toLowerCase();
-}
